Move preprocessing to corresponding pipelines
